requirements.txt
Flask==2.3.2
firebird-driver==1.0.0
Werkzeug==2.3.7
